Further migrating steps towards EntityFramework
Removed Context.Users, Context.UserDetails and Handlers.SecurityHandler. Replaced with BCrypt and direct Entity Framework usage.

Adapted the rest of the UserController and AuthenticationController API endpoints to use BCrypt and EF as well.

Also added another migration with a fix to UserName not being usable as an index because of missing maximum string length.