Introduced schemas and renamed classes and namespaces.
Things might break with this commit, so keep an eye out for any fresh bugs. However, the HaleDBModel is now renamed to UserContext and has its own default schema set to User, as per decided. Also renamed the corresponding models and namespaces so they match with what we decided on earlier. The next schema will need its own context, as in ModuleContext or NodeContext, which will contain all the relevant models for that context. Also renamed the API namespace to Controllers.